Changes to how SpyDir handles plugins
Loaded plugins are now stored with a dictionary using the type of file extension as the key. This will allow plugins to operate on multiple file types if necessary. This leads to the changes to the plugins. Plugins must now specify the types of files they operate on by maintaining a `get_ext()` function. This function must return a string containing the file extension. Multiple extensions may be included using a comma delimited string.

I broke up some of the code residing in the parse function. There were a few little bugs within that section of code that seemed most easily addressable using more modularity.